requirements.txt
fastapi
sqlalchemy
pydantic_settings
uvicorn
asyncio
asyncpg
pytest-asyncio